refactor: split appsets module
